scripts/requirements.txt
psycopg2-binary==2.9.9
jinja2==3.1.2
google-cloud-secret-manager==2.16.4
google-cloud-sql==0.3.0
google-auth==2.23.4
google-oauth2-tool==0.0.3 